[masa]: forcing all cos() evalations to be long double

+++ tests/heat_source.c
@@ -33,13 +33,13 @@
 double eval_1d_exact_t_(double x,double A_x)
 {  
   double exact_t;
-  exact_t = cos(A_x * x);
+  exact_t = cosl(A_x * x);
   return exact_t;
 }
  
 double eval_1d_source_t_(double x, double A_x, double k_0)
 {
-  double Q_T = A_x * A_x * k_0 * cos(A_x * x);
+  double Q_T = A_x * A_x * k_0 * cosl(A_x * x);
   return Q_T;
 }
 
@@ -55,7 +55,7 @@ double eval_1d_source_t_(double x, double A_x, double k_0)
 double eval_2d_exact_t_(double x,double y,double A_x,double B_y)
 {
   double exact_t;
-  exact_t = cos(A_x * x) * cos(B_y * y);
+  exact_t = cosl(A_x * x) * cosl(B_y * y);
   return exact_t;
 }
 
@@ -66,7 +66,7 @@ double eval_2d_source_t_(
   double B_y,
   double k_0)
 {
-  double Q_T = k_0 * cos(A_x * x) * cos(B_y * y) * (A_x * A_x + B_y * B_y);
+  double Q_T = k_0 * cosl(A_x * x) * cosl(B_y * y) * (A_x * A_x + B_y * B_y);
   return Q_T;
 }
 
@@ -81,7 +81,7 @@ double eval_2d_source_t_(
 double eval_3d_exact_t_(double x,double y,double z,double A_x,double B_y,double C_z)
 {
   double exact_t;
-  exact_t = cos(A_x * x) * cos(B_y * y) * cos(C_z * z);
+  exact_t = cosl(A_x * x) * cosl(B_y * y) * cosl(C_z * z);
   return exact_t;
 }
 
@@ -94,6 +94,6 @@ double eval_3d_source_t_(
   double C_z,
   double k_0)
 {
-  double Q_T = k_0 * cos(A_x * x) * cos(B_y * y) * cos(C_z * z) * (A_x * A_x + B_y * B_y + C_z * C_z);
+  double Q_T = k_0 * cosl(A_x * x) * cosl(B_y * y) * cosl(C_z * z) * (A_x * A_x + B_y * B_y + C_z * C_z);
   return Q_T;
 }
